+++ src/unpoly-migrate/network.js
@@ -247,3 +247,11 @@ function disableProgressBarIfCustomIndicator() {
 
 disableProgressBarIfCustomIndicator()
 up.on('up:framework:reset', disableProgressBarIfCustomIndicator)
+
+up.network.shouldReduceRequests = function() {
+  up.migrate('up.network.shouldReduceRequests() has been removed without replacement')
+  return false
+}
+
+up.migrate.removedProperty(up.network.config, 'badRTT')
+up.migrate.removedProperty(up.network.config, 'badDownlink')