+++ src/unpoly/classes/fragment_polling.js
@@ -3,10 +3,16 @@ const e = up.element
 up.FragmentPolling = class FragmentPolling {
 
   constructor(fragment) {
-    this.options = {}
-    this.state = 'initialized'
-    this.setFragment(fragment)
+    this.options = up.radio.pollOptions(fragment)
+
+    this.fragment = fragment
+    up.destructor(fragment, this.onFragmentDestroyed.bind(this))
+    up.fragment.onAborted(fragment, this.onFragmentAborted.bind(this))
+
+    this.state = 'initialized' // 'initialized' || 'started' || 'stopped'
     this.abortable = true
+    this.loading = false
+    this.satisfyInterval()
   }
 
   static forFragment(fragment) {
@@ -23,23 +29,26 @@ up.FragmentPolling = class FragmentPolling {
     this.stop()
   }
 
-  onFragmentAborted() {
-    if (this.abortable) {
-      this.stop()
-    }
-  }
+  start(options) {
+    Object.assign(this.options, options)
 
-  start() {
     if (this.state !== 'started') {
+      if (!up.fragment.isTargetable(this.fragment)) {
+        up.warn('[up-poll]', 'Cannot poll untargetable fragment %o', this.fragment)
+        return
+      }
+
       this.state = 'started'
-      this.scheduleReload()
+      this.ensureEventsBound()
+      this.scheduleRemainingTime()
     }
   }
 
   stop() {
     if (this.state === 'started') {
-      clearTimeout(this.reloadTimer)
+      this.clearReloadTimer()
       this.state = 'stopped'
+      this.unbindEvents?.()
     }
   }
 
@@ -54,82 +63,163 @@ up.FragmentPolling = class FragmentPolling {
     this.forceStarted = false
   }
 
-  scheduleReload(delay = this.getInterval()) {
-    this.reloadTimer = setTimeout(() => this.reload(), delay)
-  }
-
-  reload() {
-    // The setTimeout(doReload) callback might already be scheduled
-    // before the polling stopped.
-    if (this.state !== 'started') { return }
-
-    let issue = up.radio.pollIssue(this.fragment)
-    if (issue) {
-      up.puts('[up-poll]', `Will not poll: ${issue}`)
-      // Reconsider after 10 seconds at most
-      let reconsiderDisabledDelay = Math.min(10 * 1000, this.getInterval())
-      this.scheduleReload(reconsiderDisabledDelay)
-    } else {
-      this.reloadNow()
+  ensureEventsBound() {
+    if (!this.unbindEvents) {
+      this.unbindEvents = up.on('visibilitychange up:layer:opened up:layer:dismissed up:layer:accepted', this.onVisibilityChange.bind(this))
     }
   }
 
+  onVisibilityChange() {
+    if (this.isFragmentVisible()) {
+      this.scheduleRemainingTime()
+    } else {
+      // Let the existing timer play out for two reasons:
+      //
+      // 1. The fragment may become visible again before the timer runs out
+      // 2. Print the reason for not polling to the console once.
+    }
+  }
+
+  isFragmentVisible() {
+    return (this.options.ifTab   === 'any' || !document.hidden) &&
+           (this.options.ifLayer === 'any' || this.isOnFrontLayer())
+  }
+
+  clearReloadTimer() {
+    clearTimeout(this.reloadTimer)
+    this.reloadTimer = null
+  }
+
+  scheduleRemainingTime() {
+    if (!this.reloadTimer && !this.loading) {
+      this.clearReloadTimer()
+      this.reloadTimer = setTimeout(
+        this.onTimerReached.bind(this),
+        this.getRemainingDelay()
+      )
+    }
+  }
+
+  onTimerReached() {
+    this.reloadTimer = null
+    this.tryReload()
+  }
+
+  tryReload() {
+    // The setTimeout(doReload) callback might already be scheduled
+    // before the polling stopped.
+    if (this.state !== 'started') {
+      return
+    }
+
+    if (!this.isFragmentVisible()) {
+      up.puts('[up-poll]', 'Will not poll hidden fragment')
+      // (1) We don't need to re-schedule a timer here. onVisibilityChange() will do that for us.
+      // (2) Also we prefer to not have pending timers in inactive tabs to save resources.
+      return
+    }
+
+    if (up.emit(this.fragment, 'up:fragment:poll', { log: ['Polling fragment', this.fragment] }).defaultPrevented) {
+      up.puts('[up-poll]', 'User prevented up:fragment:poll event')
+      this.satisfyInterval() // Block polling for a full interval
+      this.scheduleRemainingTime() // There is no event that would re-schedule for us
+      return
+    }
+
+    this.reloadNow()
+  }
+
+  getFullDelay() {
+    return this.options.interval ?? e.numberAttr(this.fragment, 'up-interval') ?? up.radio.config.pollInterval
+  }
+
+  getRemainingDelay() {
+    return Math.max(this.getFullDelay() - this.getFragmentAge(), 0)
+  }
+
+  getFragmentAge() {
+    return new Date() - this.lastAttempt
+  }
+
+  isOnFrontLayer() {
+    this.layer ||= up.layer.get(this.fragment)
+    return this.layer?.isFront?.()
+  }
+
   reloadNow() {
+    // If we were called manually (not by a timeout), clear a scheeduled timeout to prevent concurrency.
+    // The timeout will be re-scheduled by this.onReloadSuccess() or this.onReloadFailure().
+    this.clearReloadTimer()
+
     let reloadOptions = {
       url: this.options.url,
       fail: false,
       background: true,
     }
 
-    // Prevent our own reloading from aborting ourselves.
     let oldAbortable = this.abortable
-    this.abortable = false
-    up.reload(this.fragment, reloadOptions).then(
-      this.onReloadSuccess.bind(this),
-      this.onReloadFailure.bind(this)
-    )
-    this.abortable = oldAbortable
+
+    try {
+      // Prevent our own reloading from aborting ourselves.
+      this.abortable = false
+
+      // Don't schedule timers while we're loading. onReloadSuccess() and onReloadFailure() will do that for us.
+      this.loading = true
+
+      up.reload(this.fragment, reloadOptions).then(
+        this.onReloadSuccess.bind(this),
+        this.onReloadFailure.bind(this)
+      )
+    } finally {
+      // Now that our own render pass has process abort options (this happens sync),
+      // we can resume listening to abort signals.
+      this.abortable = oldAbortable
+    }
+  }
+
+  onFragmentAborted({ newLayer }) {
+    // We temporarily set this.abortable to false while we're reloading our fragment, which also aborts our fragment.
+    if (this.abortable && !newLayer) {
+      this.stop()
+    }
   }
 
   onReloadSuccess({ fragment }) {
-    // Transfer this instance to the new fragment.
-    // We can remove this in case we don't implement forced start/stop.
+    this.loading = false
+    this.satisfyInterval()
+
     if (fragment) {
-      // No need to scheduleReload() in this branch:
-      // (1) Either the new fragment also has an [up-poll] and we have already
-      //     started in #onPollAttributeObserved().
+      // No need to scheduleRemainingTime() in this branch:
+      //
+      // (1) Either the new fragment also has an [up-poll] and we have already started in #onPollAttributeObserved().
       // (2) Or we are force-started and we will start in #onFragmentSwapped().
       this.onFragmentSwapped(fragment)
     } else {
       // The server may have opted to not send an update, e.g. if there is no fresher content.
       // In that case we try again in the next interval.
-      this.scheduleReload()
+      this.scheduleRemainingTime()
     }
-  }
-
-  onReloadFailure(reason) {
-    this.scheduleReload()
-    up.error.rethrowCritical(reason)
   }
 
   onFragmentSwapped(newFragment) {
     this.stop()
 
     if (this.forceStarted && up.fragment.matches(this.fragment, newFragment)) {
+      // Force start the new up.Polling instance for the new fragment.
       this.constructor.forFragment(newFragment).forceStart(this.options)
     }
   }
 
-  setFragment(newFragment) {
-    this.fragment = newFragment
-    up.destructor(newFragment, () => this.onFragmentDestroyed())
-    up.fragment.onAborted(newFragment, () => this.onFragmentAborted())
+  onReloadFailure(reason) {
+    this.loading = false
+    this.satisfyInterval()
+    this.scheduleRemainingTime()
+    up.error.rethrowCritical(reason)
   }
 
-
-  getInterval() {
-    let interval = this.options.interval ?? e.numberAttr(this.fragment, 'up-interval') ?? up.radio.config.pollInterval
-    return up.radio.config.stretchPollInterval(interval)
+  satisfyInterval() {
+    // This will delay the next timer scheduling for a full interval.
+    this.lastAttempt = new Date()
   }
 
 }